refactor(opensearch): removed the folder structure flaten the structure
removed the folder structure flaten the structure
