+++ python/cloudtrail-streaming-to-opensearch-serverless/.dockerignore
@@ -1,0 +1,1 @@
+cdk*
