+++ python/cloudtrail-streaming-to-opensearch-serverless/bootstrap.sh
@@ -9,9 +9,6 @@ source .env/bin/activate
 # download requirements
 .env/bin/python -m pip install -r requirements.txt --upgrade pip
 
-# Load dependency for lambda functions,
-.env/bin/python -m pip install --target lambda/build_os_client_and_bulk_ingest_logevents_handler/ -r lambda/build_os_client_and_bulk_ingest_logevents_handler/requirements.txt
-
 # load boto3
 .env/bin/python -m pip install boto3
