--- python/cloudtrail-streaming-to-opensearch-serverless/lambda/build_os_client_and_bulk_ingest_logevents_handler/requirements.txt
@@ -1,2 +1,0 @@
-
-
