--- python/cloudtrail-streaming-to-opensearch-serverless/requirements-dev.txt
@@ -1,1 +1,0 @@
-pytest==6.2.5
